requirements.txt
django-taggit
django-markup
South
